requirements.txt
alembic==1.14.0
Flask==3.1.0
Flask_Cors==5.0.0
flask_sqlalchemy==3.1.1
openai==1.57.0
pdf2image==1.17.0
Pillow==11.0.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
tiktoken==0.8.0
Werkzeug==3.1.3
